app/requirement.txt
fastapi
uvicorn[standard]
websockets
requests
websocket-client
python-dotenv